static analysis

+++ decoder.h
@@ -147,7 +147,7 @@ namespace bb::detail {
       {
          const auto [byte_index, bit_index] = div(i, 8);
          const int left_shift_amount = 7 - bit_index; // 7 is leftest bit
-         const uint8_t mask = 1ui8 << left_shift_amount;
+         const uint8_t mask = static_cast<uint8_t>(1ui8 << left_shift_amount);
          const bool binary_value = static_cast<bool>((source[byte_index] & mask) >> left_shift_amount);
          const auto replacement_color = std::bit_cast<T>(binary_value ? color1 : color0);
          target[i] = replacement_color;
@@ -366,7 +366,7 @@ constexpr auto bb::get_element_count(
    else
    {
       // Byte count is exact for everything but Indexed images
-      const int byte_count = head.bit_count / 8;
+      const int byte_count = static_cast<int>(head.bit_count) / 8;
       return byte_count / sizeof(user_type);
    }
 }
